Output info to stderr, add -0 option to separate commands on output
Now with the -s option, informational messages are prefixed with '#',
and the -0 option will change the output separator to the null-byte.

Combined with xargs -0 option, this should make it simple to customize
or parallelize commands. Example:

  $ bashdag -f dag.conf -s -0 | xargs -0 -P -L5

This would run 5 commands in parallel at a time, even if the program to be run
has newlines in it.